Rearrange packages and add subpackages
