+++ src/main/java/reposense/authorship/analyzer/FileInfoAnalyzer.java
@@ -1,4 +1,4 @@
-package reposense.authorship;
+package reposense.authorship.analyzer;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -6,11 +6,16 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.HashMap;
+import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import reposense.dataobject.Author;
-import reposense.dataobject.RepoConfiguration;
+import reposense.authorship.model.FileInfo;
+import reposense.authorship.model.FileResult;
+import reposense.authorship.model.LineInfo;
+import reposense.model.Author;
+import reposense.model.RepoConfiguration;
+import reposense.system.CommandRunner;
 import reposense.system.LogsManager;
 
 /**
@@ -20,6 +25,7 @@ public class FileInfoAnalyzer {
     private static final Logger logger = LogsManager.getLogger(FileInfoAnalyzer.class);
 
     private static final String REUSED_TAG = "//@reused";
+    private static final int AUTHOR_NAME_OFFSET = "author ".length();
 
     /**
      * Analyzes the lines of the file, given in the {@code fileInfo}, that has changed in the time period provided
@@ -33,7 +39,7 @@ public class FileInfoAnalyzer {
             return null;
         }
 
-        BlameAnalyzer.aggregateBlameAuthorInfo(config, fileInfo);
+        aggregateBlameAuthorInfo(config, fileInfo);
 
         if (config.isNeedCheckStyle()) {
             CheckStyleParser.aggregateStyleIssue(fileInfo, config.getRepoRoot());
@@ -59,6 +65,30 @@ public class FileInfoAnalyzer {
             authorContributionMap.put(author, authorContributionMap.getOrDefault(author, 0) + 1);
         }
         return new FileResult(fileInfo.getPath(), fileInfo.getLines(), authorContributionMap);
+    }
+
+    /**
+     * Sets the {@code Author} for each line in {@code fileInfo} based on the git blame analysis on the file.
+     */
+    private static void aggregateBlameAuthorInfo(RepoConfiguration config, FileInfo fileInfo) {
+        Map<String, Author> authorAliasMap = config.getAuthorAliasMap();
+
+        String blameResults = getGitBlameResult(config, fileInfo.getPath());
+        String[] blameResultLines = blameResults.split("\n");
+        int lineCount = 0;
+
+        for (String line : blameResultLines) {
+            String authorRawName = line.substring(AUTHOR_NAME_OFFSET);
+            Author author = authorAliasMap.getOrDefault(authorRawName, new Author(Author.UNKNOWN_AUTHOR_GIT_ID));
+            fileInfo.setLineAuthor(lineCount++, author);
+        }
+    }
+
+    /**
+     * Returns the analysis result from running git blame on {@code filePath}.
+     */
+    private static String getGitBlameResult(RepoConfiguration config, String filePath) {
+        return CommandRunner.blameRaw(config.getRepoRoot(), filePath, config.getSinceDate(), config.getUntilDate());
     }
 
     /**
