+++ src/main/java/reposense/system/LogsManager.java
@@ -98,6 +98,7 @@ public class LogsManager {
 
     /**
      * Creates a {@code FileHandler} for the log file.
+     *
      * @throws IOException if there are problems opening the file.
      */
     private static FileHandler createFileHandler() throws IOException {
