+++ src/test/java/reposense/parser/ArgsParserTest.java
@@ -14,7 +14,7 @@ import java.util.Optional;
 import org.junit.Assert;
 import org.junit.Test;
 
-import reposense.dataobject.CliArguments;
+import reposense.model.CliArguments;
 import reposense.util.TestUtil;
 
 public class ArgsParserTest {
