+++ src/test/java/reposense/template/GitTestTemplate.java
@@ -11,15 +11,15 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.BeforeClass;
 
-import reposense.authorship.FileInfo;
-import reposense.authorship.FileInfoAnalyzer;
-import reposense.authorship.FileInfoGenerator;
-import reposense.authorship.FileResult;
-import reposense.authorship.LineInfo;
-import reposense.dataobject.Author;
-import reposense.dataobject.RepoConfiguration;
+import reposense.authorship.analyzer.FileInfoAnalyzer;
+import reposense.authorship.extractor.FileInfoGenerator;
+import reposense.authorship.model.FileInfo;
+import reposense.authorship.model.FileResult;
+import reposense.authorship.model.LineInfo;
 import reposense.git.GitDownloader;
 import reposense.git.GitDownloaderException;
+import reposense.model.Author;
+import reposense.model.RepoConfiguration;
 import reposense.system.CommandRunner;
 import reposense.util.Constants;
 import reposense.util.FileUtil;
