+++ src/test/java/reposense/util/TestUtil.java
@@ -58,8 +58,8 @@ public class TestUtil {
     }
 
     /**
-    * Creates and returns a {@code Date} object with the specified {@code year}, {@code month}, {@code day}.
-    */
+     * Creates and returns a {@code Date} object with the specified {@code year}, {@code month}, {@code day}.
+     */
     public static Date getDate(int year, int month, int date) {
         return new Calendar
                 .Builder()
